Adding Configuration Script and Changing Makefile
Added a new configuration script, which has to be run before starting
the test. The configuration script initializes the environmental vars
for the makefiles to use. The build output and binaries are generated
into a separate directory and there is also provision for the run
results and outputs to be stores in separate directory